Fix rooms/urls + views.py
